Add files via upload
This Builds on V3. The prompt in prompts loop has been removed to be prepared to test for just one prompt. I have incorporated Exponential backing off. This can be seen in the changes made to the function in step 2. This is an example of manual back off implementation :https://platform.openai.com/docs/guides/rate-limits/error-mitigation?context=tier-free